Feature/create equipment

+++ src/main/java/com/example/myHorseServer/dto/horse/Breed.java
@@ -1,4 +1,5 @@
 package com.example.myHorseServer.dto.horse;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,8 +14,8 @@ public class Breed {
     private Integer breedId;
     private HorseBreed horseBreed;
     private double fast;
-    private Integer hungry;
-    private Integer thirst;
+    private double hungry;
+    private double thirst;
     private double appearance;
     private double value;
 }
